+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/error/MissingSourceException.java
@@ -1,8 +1,0 @@
-package uk.gov.justice.services.event.sourcing.subscription.error;
-
-public class MissingSourceException extends RuntimeException {
-
-    public MissingSourceException(final String message) {
-        super(message);
-    }
-}