+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/NewEventBufferManager.java
@@ -7,8 +7,8 @@ import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.event.buffer.core.repository.streambuffer.EventBufferEvent;
 import uk.gov.justice.services.event.buffer.core.repository.streambuffer.NewEventBufferRepository;
 import uk.gov.justice.services.event.sourcing.subscription.error.MissingPositionInStreamException;
-import uk.gov.justice.services.event.sourcing.subscription.error.MissingSourceException;
 import uk.gov.justice.services.eventsourcing.source.api.streams.MissingStreamIdException;
+import uk.gov.justice.services.eventsourcing.util.messaging.EventSourceNameCalculator;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.JsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.Metadata;
@@ -32,6 +32,9 @@ public class NewEventBufferManager {
     @Inject
     private UtcClock clock;
 
+    @Inject
+    private EventSourceNameCalculator eventSourceNameCalculator;
+
     public Optional<JsonEnvelope> getNextFromEventBuffer(
             final JsonEnvelope incomingJsonEnvelope,
             final String componentName) {
@@ -41,10 +44,7 @@ public class NewEventBufferManager {
                 format("No streamId found in event. name '%s', eventId '%s'",
                         metadata.name(),
                         metadata.id())));
-        final String source = metadata.source().orElseThrow(() -> new MissingSourceException(
-                format("No source found in event. name '%s', eventId '%s'",
-                        metadata.name(),
-                        metadata.id())));
+        final String source = eventSourceNameCalculator.getSource(incomingJsonEnvelope);
         final long currentPosition = metadata.position().orElseThrow(() -> new MissingPositionInStreamException(
                 format("No position found in event. name '%s', eventId '%s'",
                         metadata.name(),