+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/manager/BufferedEventFactoryTest.java
@@ -13,8 +13,8 @@ import static org.mockito.Mockito.when;
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.event.buffer.core.repository.streambuffer.EventBufferEvent;
 import uk.gov.justice.services.event.sourcing.subscription.error.MissingPositionInStreamException;
-import uk.gov.justice.services.event.sourcing.subscription.error.MissingSourceException;
 import uk.gov.justice.services.eventsourcing.source.api.streams.MissingStreamIdException;
+import uk.gov.justice.services.eventsourcing.util.messaging.EventSourceNameCalculator;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.JsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.Metadata;
@@ -33,6 +33,9 @@ public class BufferedEventFactoryTest {
 
     @Mock
     private JsonObjectEnvelopeConverter jsonObjectEnvelopeConverter;
+
+    @Mock
+    private EventSourceNameCalculator eventSourceNameCalculator;
 
     @InjectMocks
     private BufferedEventFactory bufferedEventFactory;
@@ -53,7 +56,7 @@ public class BufferedEventFactoryTest {
         when(jsonEnvelope.metadata()).thenReturn(metadata);
         when(metadata.streamId()).thenReturn(of(streamId));
         when(metadata.position()).thenReturn(of(positionInStream));
-        when(metadata.source()).thenReturn(of(source));
+        when(eventSourceNameCalculator.getSource(jsonEnvelope)).thenReturn(source);
         when(jsonObjectEnvelopeConverter.asJsonString(jsonEnvelope)).thenReturn(jsonEnvelopeAsJsonString);
 
         final EventBufferEvent eventBufferEvent = bufferedEventFactory.createFrom(jsonEnvelope, componentName, bufferedAt);
@@ -90,33 +93,6 @@ public class BufferedEventFactoryTest {
     }
 
     @Test
-    public void shouldThrowMissingSourceExceptionIfNoStreamIdFoundInJsonEnvelope() throws Exception {
-
-        final UUID eventId = fromString("6418d3c2-96e4-410e-81a2-aa3c1f82e9f2");
-        final UUID streamId = randomUUID();
-        final String name = "event-name";
-        final String componentName = "some-component-name";
-        final String jsonEnvelopeAsJsonString = "json-envelope-as-json-string";
-        final ZonedDateTime bufferedAt = new UtcClock().now();
-
-        final JsonEnvelope jsonEnvelope = mock(JsonEnvelope.class);
-        final long positionInStream = 23L;
-        final Metadata metadata = mock(Metadata.class);
-
-        when(jsonEnvelope.metadata()).thenReturn(metadata);
-        when(metadata.id()).thenReturn(eventId);
-        when(metadata.name()).thenReturn(name);
-        when(metadata.streamId()).thenReturn(of(streamId));
-        when(metadata.source()).thenReturn(empty());
-
-        final MissingSourceException missingSourceException = assertThrows(
-                MissingSourceException.class,
-                () -> bufferedEventFactory.createFrom(jsonEnvelope, componentName, bufferedAt));
-
-        assertThat(missingSourceException.getMessage(), is("No source found in event: name 'event-name', eventId '6418d3c2-96e4-410e-81a2-aa3c1f82e9f2'"));
-    }
-
-    @Test
     public void shouldThrowMissingPositionInStreamExceptionIfNoStreamIdFoundInJsonEnvelope() throws Exception {
 
         final UUID eventId = fromString("008eabd8-75b8-445b-8a8a-4fd218448e8a");
@@ -135,7 +111,7 @@ public class BufferedEventFactoryTest {
         when(metadata.name()).thenReturn(name);
         when(metadata.streamId()).thenReturn(of(streamId));
         when(metadata.position()).thenReturn(empty());
-        when(metadata.source()).thenReturn(of(source));
+        when(eventSourceNameCalculator.getSource(jsonEnvelope)).thenReturn(source);
 
         final MissingPositionInStreamException missingPoisitionInStreamException = assertThrows(
                 MissingPositionInStreamException.class,